refactor: Remove test workflows and consolidate to `abominable_INK_.json`
This commit removes all test workflows and leaves only `abominable_INK_.json`
as the main workflow. This workflow will be used as a base for all future
workflows in the "Abominable Workflows", which will specialize in generating
different image styles.
For example, there will be a workflow called `abominable_POLAROID_.json` that
will generate images that resemble old photos taken with a Polaroid camera.