feat: Add classic_abominable_spaghetti.json workflow
This commit adds the `classic_abominable_spaghetti.json` workflow,
which is a modernized version of the previous `v2` workflow. It
features a new layout with frequently modified nodes in the center
and less frequently modified nodes on the periphery. This workflow
will serve as a base for the node distribution of other workflows
in Abominable Workflows v3.