feat: Add abominable_grid image examples
Added two versions of `abominable_grid` to the /demo_images directory.
This image showcases the capabilities of the Abominable Workflows.
A high-resolution PNG (16MB) and a compressed JPG (1.3MB) are provided
to offer different options for viewing and usage.

Note that this deviates slightly from the one-image-per-file convention
of the `demo_images` directory, but is justified given the demonstrative
nature of this grid.